fix null in filter problem
