+++ msgvis/static/grouper.less
@@ -601,7 +601,7 @@ body {
 #filter-popup {
     position: absolute;
     top: @top-row-height + @bottom-row-height - @stat-box-height + 30;
-    left: 68%;
+    left: 63%;
 
     z-index: 1000;
